testing rdu, stop internal multi-threadding and refine the tsearch(3) call
The purpose of the internal multi-threadding approach was to issue
statat(2) to get the inode number. Now aufs (in kernel space) implements
the interface which returns the multiple inode numbers in a single
ioctl.
Throw away the multi-threadding approach. Additionally, the scheduling
on pipe does not work well in some older kernel versions.

While I don't like the interface of tsearch(3) family, it seems fast.
Let's keep on going.

Signed-off-by: J. R. Okajima <hooanon05@yahoo.co.jp>