aufs: compat_ioctl, make a room for a pointer
(A commit in a series of supporting 32bit emulation under 64bit kernel.
While every commit is git-bisect-able, you shoule read all commits in
the series since a single commit may have less meaning.)

In order to make it compatible, make a room for a pointer and always
handle it as 64bit size.

Signed-off-by: J. R. Okajima <hooanon05@yahoo.co.jp>