bugfix, O_CLOEXEC for the plink maintenance mode
/sbin/mount.aufs (and others) puts aufs into the plink maintenance mode
via /proc/fs/aufs/plink_maint which make many other operations to return
error or block.
During in this mode, /sbin/mount.aufs exec(2) the original mount(8).
If mount(8) is not statically linked, it may mmap(2) ld.so (and
others). And if ld.so is inside of the target aufs, then aufs mmap(2)
blocks, ie. deadlock.
To address this problem, specify O_CLOEXEC for /proc/fs/aufs/plink_maint
which makes aufs to exit the plink maintenance mode.

Reported-by: Marco Clocchiatti <ziapannocchia@gmail.com>
Signed-off-by: J. R. Okajima <hooanon05@yahoo.co.jp>