patch to auroot
Hi,

I'm submitting a patch to auroot (as shipped with Ubuntu 11.10, appears
the same as in git).  I had a lot of trouble understanding where the
overlapping branch was in the provided script.  I'm guessing an old
version of aufs did not fully check bind mounts and so the script used
to work.

There are only two basic changes:
1) A temporary directory is mounted using tmpfs, so there are no
overlapping branches.
2) The read-only bind out is removed.  Presumably, this was a hack
around the old overlap check and is also unnecessary, as aufs will cover
this mount with a rw mount.

Thanks,
Jayen